GermanWordDaily/requirements.txt
plyer
requests
google-generativeai
